Uploaded .RAR, .ZIP and .7Z File for C# Demo Code
Extract this to a folder of your choosing and open the C# .sln file to edit the code. Alternatively, if you are only interested in running the program, a .exe file can be loacated in the .bin file.

Only one of the uploaded files is required for running. Choose whichever format works best for you.